separate var_registry.*pp completely from variable.hpp
