+++ src/crab/ebpf_domain.cpp
@@ -14,6 +14,7 @@
 #include "config.hpp"
 #include "crab/array_domain.hpp"
 #include "crab/ebpf_domain.hpp"
+#include "crab/var_registry.hpp"
 #include "string_constraints.hpp"
 
 namespace prevail {
