+++ src/crab_utils/stats.cpp
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 #include "stats.hpp"
 
+#include <iostream>
 #include <optional>
 #ifdef _WIN32
 #include <windows.h>
