+++ src/crab_verifier.cpp
@@ -13,6 +13,7 @@
 #include "asm_syntax.hpp"
 #include "crab/ebpf_domain.hpp"
 #include "crab/fwd_analyzer.hpp"
+#include "crab/var_registry.hpp"
 #include "crab_utils/lazy_allocator.hpp"
 #include "crab_verifier.hpp"
 #include "string_constraints.hpp"
